ENGR00112182 Enable basic telephony feature with HuaWei EM750 module
1 kernel_imx:
  * Fix UART3 IOMUX configuration error (found also in GPS integration)
2 vendor/fsl
  * Add "BOARD_HAVE_MODEM" and other macros in BoardConfig.mk
  * Add rild parameter in init.rc
  * Add new service for "pppd_gprs"
3 system/core
  * Add "/dev/ttymxc2" (i.e. UART3) into permission table in
devices.c so that it can be accessed by rild
  * Add "pppd_gprs" property permission in property_service.c so that
rild can request to start/stop "pppd_gprs" service
  * Change Android.mk in rootdir to install PPP relateted files (e.g.
ppp/ip-up)
  * Add pppd parameter (for EM750) into init.gprs-pppd script
  * Set "net.dns1" property in ip-up script (this is temporary
workaround)
4 external/ppp
  * Add "-fno-short-enums" to force not using short enum since kernel
PPP driver will treat enum as 4 byte
  * Add one patch to enable "-replacedefaultroute" option in pppd.
With this option, after connecting with mobile, the new ppp0 link can
be added to route table as the default route (i.e. replace the old
default route if there is already one)
  * Set remote ip address same as local IP address when remote IP
address can't be determined in IPCP negotation. PPPD default set a
dummy address as remote IP address. This will cause the "gateway" of
default route in route table can't be reached.
  * log format string "%I" can't be recognized by Android log utility.
So change it to "%s" with a local ip_ntoa() call
5. hardware/ril
  * Start pppd when get response for PDP activation
  * Poll ppp link status until it's running or fail after timeout
  * Add "-u" option for libreference-ril so that it will only
start initialization when /dev/ttyUSB0 is available (that device will
be created when EM750 is power up and finish USB enumaration with
i.MX51)
  * Set termio of UART3 to configure parameters like baud rate