ENGR00136152 add infineon 3G modem special ril code.
add a vendor type AMAZON, since this modem called Amazon.

Fix for issue where connection does not start and the APN
is not displayed. During startup, RILD calls the function
requestRadioPower(off) as a means to reset the modem. It
also uses this function to enter airplane mode. The current
implementation sends the AT+CFUN=0 command, which causes
the Amazon module to shutdown the USB interface.
A temporary fix for this problem is to issue AT+CFUN=4
instead. This powers down the radio, but does not disable
the USB interface.

A permanent fix should use GPIO lines to reset the modem and
manage the modem power states more carefully.

The HuaWei module integration uses the radio interface layer
daemon (RILD) to establish a data connection by sending AT
commands over /dev/ttyUSB3. Once the connection is established,
it uses a shell script to start the PPP daemon on /dev/ttyUSB0.
This process fails with the Amazon module, because it times out
(1 sec) waiting for the PPP connection.
The result is PPPD exiting with error 17, EXIT_LOOPBACK.

The fix for this problem is to have the PPPD dial the connection.
RILD continues to manage all setup and configuration of the modem,
but no longer issues the dial command

ref: http://wiki.freescale.net/display/MADCPOIMX/Infineon+3G+Mobile

Signed-off-by: Zhang Jiejing <jiejing.zhang@freescale.com>