ENGR00137739 ril: Hot plug supported 3G modem function.
If none of data or AT command interface is given in
init.rc, try to use vendorID & productID to figure
out which port should be use according a supported
modem table.

Add a watching thread, if the supported 3G modem
is plug in or plug out, will restart rild to fit
new modem or enter no modem state.

To support runtime detect 3G modem type and hotplug,
needs replace macro define to a global var of what type
of modem we are using.

Signed-off-by: Zhang Jiejing <jiejing.zhang@freescale.com>